+++ 1.7.10/src/main/java/net/geforcemods/securitycraft/blocks/BlockLaserBlock.java
@@ -273,9 +273,9 @@ public class BlockLaserBlock extends BlockContainer {
     @SideOnly(Side.CLIENT)
     public void randomDisplayTick(World par1World, int par2, int par3, int par4, Random par5Random){
     	if(par1World.getBlockMetadata(par2, par3, par4) == 2){
-    		double d0 = (double)((float)par2 + 0.5F) + (double)(par5Random.nextFloat() - 0.5F) * 0.2D;
-    		double d1 = (double)((float)par3 + 0.7F) + (double)(par5Random.nextFloat() - 0.5F) * 0.2D;
-    		double d2 = (double)((float)par4 + 0.5F) + (double)(par5Random.nextFloat() - 0.5F) * 0.2D;
+    		double d0 = par2 + 0.5F + (par5Random.nextFloat() - 0.5F) * 0.2D;
+    		double d1 = par3 + 0.7F + (par5Random.nextFloat() - 0.5F) * 0.2D;
+    		double d2 = par4 + 0.5F + (par5Random.nextFloat() - 0.5F) * 0.2D;
     		double d3 = 0.2199999988079071D;
     		double d4 = 0.27000001072883606D;
