+++ 1.7.10/src/main/java/net/geforcemods/securitycraft/blocks/BlockLogger.java
@@ -55,7 +55,7 @@ public class BlockLogger extends BlockOwnable {
     {
     	super.onBlockPlacedBy(p_149689_1_, p_149689_2_, p_149689_3_, p_149689_4_, p_149689_5_, p_149689_6_);
     	
-        int l = MathHelper.floor_double((double)(p_149689_5_.rotationYaw * 4.0F / 360.0F) + 0.5D) & 3;
+        int l = MathHelper.floor_double(p_149689_5_.rotationYaw * 4.0F / 360.0F + 0.5D) & 3;
 
         if (l == 0)
         {
