+++ 1.7.10/src/main/java/net/geforcemods/securitycraft/network/packets/PacketCPlaySoundAtPos.java
@@ -1,13 +1,13 @@
 package net.geforcemods.securitycraft.network.packets;
 
-import net.minecraft.client.Minecraft;
-import io.netty.buffer.ByteBuf;
 import cpw.mods.fml.common.network.ByteBufUtils;
 import cpw.mods.fml.common.network.simpleimpl.IMessage;
 import cpw.mods.fml.common.network.simpleimpl.IMessageHandler;
 import cpw.mods.fml.common.network.simpleimpl.MessageContext;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
+import io.netty.buffer.ByteBuf;
+import net.minecraft.client.Minecraft;
 
 public class PacketCPlaySoundAtPos implements IMessage{
 	
@@ -55,7 +55,7 @@ public static class Handler extends PacketHelper implements IMessageHandler<Pack
 
 	@SideOnly(Side.CLIENT)
 	public IMessage onMessage(PacketCPlaySoundAtPos message, MessageContext ctx) {
-		Minecraft.getMinecraft().theWorld.playSound((double) message.x,(double) message.y,(double) message.z, message.sound, (float) message.volume, 1.0F, true);
+		Minecraft.getMinecraft().theWorld.playSound(message.x,message.y,message.z, message.sound, (float) message.volume, 1.0F, true);
 		return null;
 	}
